requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch
torchvision
torchaudio
byaldi
transformers==4.46.3
peft==0.13.2
accelerate
qwen-vl-utils
gradio
pandas
bitsandbytes
Pillow